requirements.txt
Flask-SocketIO==4.3.1
python-engineio==3.13.2
python-socketio==4.6.0
Flask==2.0.3
Werkzeug==2.0.3
opencv-contrib-python
pillow
tensorflow-hub
numpy==1.24.0
matplotlib
WooCommerce
opencv-python
torch
nltk